feat: 优化swagger 3.0

+++ mcdull-support/mcdull-uac/mcdull-uac-provider/src/main/java/io/gitee/dqcer/mcdull/uac/provider/model/dto/RolePermissionInsertDTO.java
@@ -8,5 +8,5 @@ import java.util.List;
 @Data
 public class RolePermissionInsertDTO implements DTO {
 
-    private List<Integer> menuIdList;
+    private List<Long> menuIdList;
 }
