+++ mcdull-support/mcdull-uac/mcdull-uac-provider/src/main/java/io/gitee/dqcer/mcdull/uac/provider/web/controller/RoleController.java
@@ -48,12 +48,12 @@ public class RoleController {
      * @return {@link Result<Integer> 返回新增主键}
      */
     @PostMapping("insert")
-    public Result<Integer> insert(@RequestBody @Validated RoleInsertDTO dto){
+    public Result<Long> insert(@RequestBody @Validated RoleInsertDTO dto){
         return Result.success(roleService.insert(dto));
     }
 
     @PutMapping("{id}/update")
-    public Result<Boolean> update(@PathVariable("id") Integer id, @RequestBody @Validated RoleUpdateDTO dto){
+    public Result<Boolean> update(@PathVariable("id") Long id, @RequestBody @Validated RoleUpdateDTO dto){
         return Result.success(roleService.update(id, dto));
     }
 
@@ -64,7 +64,7 @@ public class RoleController {
      * @return {@link Result<Integer>}
      */
     @PutMapping("{id}/status")
-    public Result<Boolean> toggleStatus(@PathVariable("id") Integer id, @RequestBody ReasonDTO dto){
+    public Result<Boolean> toggleStatus(@PathVariable("id") Long id, @RequestBody ReasonDTO dto){
         return Result.success(roleService.toggleStatus(id, dto));
     }
 
@@ -75,12 +75,12 @@ public class RoleController {
      * @return {@link Result<Integer>}
      */
     @DeleteMapping("{id}")
-    public Result<Boolean> delete(@PathVariable("id") Integer id, @Validated ReasonDTO dto){
+    public Result<Boolean> delete(@PathVariable("id") Long id, @Validated ReasonDTO dto){
         return Result.success(roleService.delete(id, dto));
     }
 
     @PutMapping("{id}/permission")
-    public Result<Boolean> insertPermission(@PathVariable("id") Integer id, @RequestBody RolePermissionInsertDTO dto){
+    public Result<Boolean> insertPermission(@PathVariable("id") Long id, @RequestBody RolePermissionInsertDTO dto){
         return Result.success(roleService.insertPermission(id, dto));
     }
