+++ mcdull-support/mcdull-uac/mcdull-uac-provider/src/main/java/io/gitee/dqcer/mcdull/uac/provider/web/service/impl/RoleMenuServiceImpl.java
@@ -19,7 +19,7 @@ import java.util.Map;
 public class RoleMenuServiceImpl extends BasicServiceImpl<IRoleMenuRepository> implements IRoleMenuService {
 
     @Override
-    public Map<Integer, List<Integer>> getMenuIdListMap(List<Integer> roleIdList) {
+    public Map<Long, List<Long>> getMenuIdListMap(List<Long> roleIdList) {
         if (CollUtil.isNotEmpty(roleIdList)) {
             return baseRepository.menuIdListMap(roleIdList);
         }
@@ -27,7 +27,7 @@ public class RoleMenuServiceImpl extends BasicServiceImpl<IRoleMenuRepository> i
     }
 
     @Override
-    public boolean deleteAndInsert(Integer roleId, List<Integer> menuIdList) {
+    public boolean deleteAndInsert(Long roleId, List<Long> menuIdList) {
         List<RoleMenuEntity> list = baseRepository.listByRoleId(roleId);
         if (CollUtil.isNotEmpty(list)) {
             baseRepository.removeBatchByIds(list);
@@ -39,7 +39,7 @@ public class RoleMenuServiceImpl extends BasicServiceImpl<IRoleMenuRepository> i
     }
 
     @Override
-    public Map<Integer, List<Integer>> getRoleIdMap(List<Integer> menuIdList) {
+    public Map<Long, List<Long>> getRoleIdMap(List<Long> menuIdList) {
         if (CollUtil.isNotEmpty(menuIdList)) {
             return baseRepository.listByMenuIdList(menuIdList);
         }
